rewrite to Double weights, g(20000, 200), s=1.5 or 3?, experiment with Double/double

+++ src/ShortestPathFinder.java
@@ -1,7 +1,7 @@
 import java.util.List;
 
 interface ShortestPathFinder {
-    record Result (List<Integer> distances, List<Integer> predecessors) {}
+    record Result (List<Double> distances, List<Integer> predecessors) {}
 
     Result findShortestPaths(Graph g, int source);
 }
